Updated modules, changed source type

+++ src/lib/math.js
@@ -14,7 +14,7 @@ const math = create(
   {
     matrix: 'Array',
     number: 'BigNumber',
-    precision: 32
+    precision: 15 // Hardcoded!
   }
 )
